Feat: you  now save  favourites
